ui(info): removed footer and added info dropdown to nav
